git-svn-id: https://svn.cebitec.uni-bielefeld.de/svn/mgx@1575 2f99df6c-41e9-6383-b5e2-de53e728a4e2

+++ MGX-maven-ejb/src/main/java/de/cebitec/mgx/model/dao/deleteworkers/DeleteReference.java
@@ -58,19 +58,21 @@ public class DeleteReference extends TaskI {
                 }
             }
             setStatus(TaskI.State.PROCESSING, "Deleting reference " + refName);
-            File f = new File(fileName);
-            if (f.exists()) {
-                f.delete();
+            if (fileName != null) {
+                File f = new File(fileName);
+                if (f.exists()) {
+                    f.delete();
+                }
             }
 
             try (PreparedStatement stmt = conn.prepareStatement("DELETE FROM region WHERE ref_id=?")) {
                 stmt.setLong(1, id);
-                stmt.execute();
+                stmt.executeUpdate();
             }
 
             try (PreparedStatement stmt = conn.prepareStatement("DELETE FROM reference WHERE id=?")) {
                 stmt.setLong(1, id);
-                stmt.execute();
+                stmt.executeUpdate();
             }
             setStatus(TaskI.State.FINISHED, refName + " deleted");
         } catch (Exception e) {
